+++ package.json
@@ -52,7 +52,8 @@
     "unified": "^8.4.0",
     "unist-util-flatmap": "^1.0.0",
     "unist-util-map": "^3.1.2",
-    "unist-util-select": "^4.0.0"
+    "unist-util-select": "^4.0.0",
+    "unist-util-visit": "^4.1.1"
   },
   "peerDependencies": {
     "@readme/variable": "^15.1.0",
