+++ processor/transform/index.js
@@ -1,0 +1,1 @@
+export { default as singleCodeTabsTransformer } from './single-code-tabs';
